monitor connection table for active_connections
[1] introduced a new key "active_connections" in status column
to store ovsdb client's ip for each active connection by ovsdb-server.

In networking-ovn, we watch for active_connections updates, fetch the
ip address of the remote client and then iterate through all chassis
and mark it as acive if the ip belongs to it. Then this list of
active chassis is notified to plugin for updating neutron DB.

[1] https://github.com/venkataanil/ovs/commit/f4582d2c7cc8349479294735306831315c8484ad